+++ application/config/autoload.php
@@ -132,5 +132,4 @@ $autoload['language'] = array();
 |
 |	$autoload['model'] = array('first_model' => 'first');
 */
-$autoload['model'] = array('app');
-$autoload['model'] = array('app', 'tasks');
+$autoload['model'] = array('app', 'tasks', 'task');